Add Access to ActionController

+++ app/Http/Controllers/ActionController.php
@@ -33,7 +33,10 @@ class ActionController extends Controller
     public function unAssignTeam(Request $request){
 
         $user = User::findOrfail($request->user_id);
-        $team = Team::findOrfail($request->team_id);
+        $team = Team::findOrfail($request->team_id)->where('owner',\Auth::id())->first();
+        if(!$team){
+            return response()->json('You can not un assign users to this team, because you are not owner', 200);
+        }
         $team->users()->detach($user);
 
         return response()->json('User Successfully assigned to Team', 200);
@@ -43,6 +46,7 @@ class ActionController extends Controller
     public function unAssignRole(Request $request){
         $user = User::findOrfail($request->user_id);
         $role = Role::findOrfail($request->role_id);
+
         $role->users()->detach($user);
 
         return response()->json('User Successfully assigned to Team', 200);
@@ -50,7 +54,10 @@ class ActionController extends Controller
 
     public function setOwner(Request $request){
         $user = User::findOrfail($request->user_id);
-        $team =Team::findOrfail($request->team_id);
+        $team = Team::findOrfail($request->team_id)->where('owner',\Auth::id())->first();
+        if(!$team){
+            return response()->json('You can not set owner to this team, because it is not your team', 200);
+        }
         $team->owner = $user->id;
         $team->save();
